test: move BLS scheme restore regression test to bls_tests
The v19_boundary_validation_failure_restores_bls_scheme regression
test exercises BLS-scheme bookkeeping after a failed block validation
at the V19 activation boundary, so it belongs alongside the other BLS
tests in src/test/bls_tests.cpp rather than the deterministicmns
suite.
